requirements.txt
flask==3.1.2
argon2-cffi==25.1.0